+++ docs/CHANGELOG.rst
@@ -15,6 +15,7 @@ Fix
 - Add default value for ``FLOW_PROCESSES_ALLOW_LIST`` and
   ``FLOW_PROCESSES_IGNORE_LIST`` in case of missing settings.
 - User defined slug must not be changed
+- Raise ``ValidationError`` when trying to get non-existent ``AnnotationField``
 
 Added
 -----